Fix all missing NSSecureCoding support.
We should favor NSSecureCoding over NSCoding in eDO itself to offend ourselves from substitution attack, although we haven't currently enabled secureCoding when serializing requests and responses. This is because NSException is not NSSecureCoding compliant, and we use it to propagate the exceptions when making remote invocations. We will need to supply our own EDOException before we can enable secureCoding but now at least we make everything else possible to be NSSecureCoding compliant.

PiperOrigin-RevId: 217928038