Unwrap an object from EDOClientService.
Unwrapping a remote object shouldn't require the service to run on the current queue. It finds the service from the current running queue so it can check if the remote object already belongs to it to be unwrapped. If there is no service or it doesn't match but the remote, it then talks to the service the remote object belongs to in the same process to check its availability.

PiperOrigin-RevId: 219534828