Remove the executor dependency on the channels.
The executor should only handle the requests and run the while loop, and not deal with the channels. The executor will continue to run the while loop and the caller is responsible to close the queue in the handler to exit the while loop. Whether the executor is running, it handles any requests in the execution queue synchronously. Encoding and decoding requests and responses, sending and receiving NSData's will be moved out of the executor logic later.

PiperOrigin-RevId: 221161894