Create a temporary service when sending a remote request if needed.
If there is no service associated with the current running queue, a temporary service will be created during the course of invocation. The temporary service will be gone after the invocation so any request for this service will fail afterwards. This can be useful when sending a remote invocation on a dispatch queue that doesn't set up a service.

PiperOrigin-RevId: 221315921