Create the listen socket lazily if the object or port is not given.
The temporary service is created every time when there is no service bound to the current running queue, this results in a lot of dummy sockets created and deleted. This change delays to create the listen socket until it is being used by the service: either given a root object to serve or wrap a local object.

PiperOrigin-RevId: 222152421