Added EDOHostNamingServiceObject to provide service ports information.
This special service object starts an EDOHostService using it as the root object. Through remote invocation to this object, users will be able to fetch EDOHostService instances information that are running.
Added a new property of serviceName to EDOHostService instance. This can help recognize service instance in the client side.

PiperOrigin-RevId: 226866810