Resolve EDOObject to the same ones for the same underlying objects in the arguments.
When a remote object is returned from the remote host, EDO checks if there is already one used in the client, and if there is one, then the same one will be used to assure the pointer equality. This change fixes the case when a remote object is passed in the argument.

PiperOrigin-RevId: 228604194