Forward the -description of EDOObject.
The default implementation doesn't have anything but only prints out the class name and the local memory address. When EDOObject is used, often times the caller would expect the underlying object's description so forwarding -description is an expected behavior. If the caller decides to inspect the actual object, it can use objc_getClass or alike to retrieve those information.

PiperOrigin-RevId: 229421071