Remove the unneeded test and rephrase the comment.
We don't need to verify that the addresses aren't the same for different objects, it's not actually guaranteed. For example, the object can be re-allocated on the same memory as the object that was just released. We only need to make sure the objects are the same if the underlying objects are the same.

PiperOrigin-RevId: 229579399