Clean up the executor after the test finishes.
The message queue needs to be closed so the executor will finish running and the queue will be released. A few flakiness was seen on the real devices to report the object resurrection crash. The actual cause is still unknown, but it's possible that this comes from an unfinished dispatch queue. This change is one attempt to fix it.

PiperOrigin-RevId: 237919010