Decouple the channel with the host port.
The EDOHostPort is used as the key for the channel pool, as well as a destination to set up a connection. It doesn't need to couple with the channel. The channel itself has an implicit destination that can't be changed once built via socket but the channel pool can key it in any way it wants to.

PiperOrigin-RevId: 258095181