Close socket at dealloc.
The device channel may release the socket and closing the channel would not be able to tell if the socket is still being help by itself, so we need to close the socket at dealloc after releasing the socket, and let the dispatch_io not to handle it at all. The cost is that if there is on-going traffic, they may get dropped.

PiperOrigin-RevId: 259063019