Use EDOHostPort to request remote objects.
EDOHostPort is used as an opaque object to connect to a channel, the client service doesn't need to expose the internal details in order to work with. This will also abstract how the client can connect to a host service, making it easier in the future to develop new ways to connect without any changes on the client APIs.

PiperOrigin-RevId: 259639501