Close file descriptors in the cleanup handler.
The dispatch_io only relinquishes the control of the underlying file descriptor after the cleanup handler is invoked. Closing the file descriptor before it is called can cause the race condition or crash. Therefore, once the EDOSocketChannel takes over the socket, it can't easily release it. Since we don't use -releaseSocket at all, it is now also removed.

PiperOrigin-RevId: 260045543