Add the multiplexer to forward channels.
The multiplexer is a reverse proxy that forwards the channel. It starts to wait on the incoming connections from another host, and instructs the other end to connect to the host that the current process is not able to connect directly. For example, the client on the real devices won't be able to connect to the host on Mac via USB.

PiperOrigin-RevId: 260828746