Add the channel forwarder to serve multiplexers.
The forwarder connects to the multiplexer and listens to connect to the host given by the EDOHostPort. The forwarder needs to understand the port received from the multiplexer in order to establish the connection. Once the connection is established, the forwarder starts to forward data between the two ends and no other redirection can be performs until it stops and starts again.

PiperOrigin-RevId: 261394712