Add the device forwarder manager.
The helper manager class manages a set of forwarders connecting to the device's multiplexer. The manager attempts to connect to the device's multiplexer and manages the established forwarders. The manager will set up two forwarder in the beginning so it can handle two concurrent connections. Once the forwarder starts to forwarding, the manager will attempt to establish another one, if it is not exceeding the limit, so the multiplexer can take concurrent requests. If the multiplexer closes, the manager will invalidate all the forwarders automatically.

PiperOrigin-RevId: 265967350