Updates message queue management in EDOExecutor.
EDOExecutor holds a list of EDOBlockingQueues to handle the execution of incoming eDO invocations. EDOBlockingQueues are created and released in last-in-first-out manner. Each of them serves an outgoing request while waiting for the corresponding response. When eDO invocation is attached to the EDOExecutor, the latter checks its top-stack EDOBlockingQueue - If it's closed, then using dispatch_async to attach the execution instead of message queue.

We have caught an issue regarding to this behavior in nested call, such that an outgoing request (whose serving EDOBlockingQueue is in lower-level of the stack) relies on an incoming request to complete, which is scheduled in dispatch_async since top-stack EDOBlockingQueue is closed but not popped out yet. Then it causes a deadlock.

To fix this issue, EDOExecutor should make sure all EDOBlockingQueue in the stack are closed before dispatching the execution.

PiperOrigin-RevId: 271613455