Decouple EDOExecutor and EDOHostService.
EDOExecutor currently does two things: 1) task management for a single thread. 2) get the handler for a EDOServiceRequest and invoke the handler. The latter relies on the logic in EDOHostService(+Handlers), which should belong to EDOHostService and adds complexity to EDOExecutor.

We can make EDOExecutor more general by accepting an executing block instead of EDOServiceRequest, and move the request handling logic back to EDOHostService. It also removes the ambiguous 'context' argument.

PiperOrigin-RevId: 280333393