Populate EDOExecutor's handleBlock: error to client.
When process A creates an anonymous host (temporary host service) and sends an object to process B, the object at process B is valid only within the scope of initial invocation made from process A. Once that invocation completes, the anonymous host will be released and no longer accept requests. If process B holds the remote object and uses it after the completion of initial invocation, the host may still have chance to process the request if it haven't been terminated. For now, it will throw an exception and crashes the host.

However, this essentially is a client-side misusage of the API. So we should send the error back to client and throw exception there. Currently, we only do it for bidirectional requests (the request that requires response), since one-way request terminates the connection right after the request being sent.

PiperOrigin-RevId: 281170599