Use lightweight generics to retrieve root objects.
In Swift, when casting, it does runtime type checks. This can fail if the object is a remote proxy (aka EDOObject). This can be bypassed by using the lightweight generics where the conversion happens in the objc level. If we provide the type to ObjC runtime at calling, the type checking happens at ObjC level, where it calls objc methods to validate, i.e. (isKindOfClass: or conformsToProtocol:); those will be then forwarded remotely and then can be successfully validated. This way, in Swift the user can directly fetch the root object through protocols.

PiperOrigin-RevId: 307746626