Fix eDO TSAN issue caused by __block variable.
When __block variable lives beyond the declarant scope and is written asynchronously in the block, there is a small chance (< 1/1000) that TSAN complains [read and write happens unsynchronized](https://screenshot.googleplex.com/6WfDaySAK2moAi4). The read, which is complained by the TSAN, implicitly added by ARC at the end of the scope through a release call.

As a workaround, we can move the `__block` declaration, as well as the blocks that use it, into the serial queue that runs the block. In that case, the implicit read on `__block` variables will happen synchronized.

PiperOrigin-RevId: 340325470