Adds one more test case for deadlock fix.
The cache for temporary service is not immediately released after the outermost eDO invocation. Instead, it is released when autorelease pool ends, such as the end of dispatch_sync/async. So adding a test case to imply that.

At the same time, fix an issue that will fail the new test case, and cause retain cycle of host service -> handler -> host service.

PiperOrigin-RevId: 361217991