Avoid always creating listening port on the temporary service.
This is a bug introduced by the last temporary service deadlock fix. When temporary service gets a response associated with a EDOObject return value, it calls `[EDOClientService -unwrappedObjectFromObject:]`. Currently this invocation will creating a listening port which is not necessary.

To fix this, we move the lazy-load of the listen socket to where it's actually needed, i.e. boxing objects. Other access to the host port won't trigger lazy-load any more.

PiperOrigin-RevId: 362571601