eDO won't block the object encoding if the object is passed by value.
This is useful when developers want to block a class which has subclasses that are value type. In this case, eDO should bypass codable subclasses since they are not passed by reference.

As an extreme use case, developers can block NSObject to disallow pass by reference at all, but still send value types across the processes.

PiperOrigin-RevId: 426274648