Add API to exclude types from eDO blocklist.
By adding a type into eDO blocklist, all of its subclasses are blocked in eDO remote calls as well. This is intended behavior, but in some scenario, people want to perform "allowlist" manner for eDO remote calls, e.g., in EarlGrey test.

The new API allows people to block NSObject while adding a list of types to form an allowlist.

PiperOrigin-RevId: 453599300