Simplify EDOObjectAliveMessage.
The original EDOObjectAliveRequest/Response takes an EDOObject and verifies if it's cached by the target server, then returns itself on success. The EDOObject parameter is redundant in both request and response, because for request it only requires `port` and `remoteAddress`, and for response it only requires a BOOL.

The simplification of the request and response will reduce the size of the network message, as well as removing the redundant encoding/decoding of the EDOObject.

PiperOrigin-RevId: 462198018