Automatically forward `isKindOfClass:` for NSError.
This attempts to automatically forward the class comparison when a class with the same name can be found in the remote process.

The main reason for this change is that Xcode 16.3 seems to introduce a new behavior in Swift tests. In particular, an error that is thrown will be captured by a backtracing function that force casts it to a Swift error. This cast will fail and crash the test when the receiver is an `EDOObject`, which cannot be casted to a `Swift.Error`.

Forwarding the type check with the proper class fixes the issue.

PiperOrigin-RevId: 783838546